// app/src/api/model/exportFileRequestBodyFormat.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Economicon API
 * OpenAPI spec version: 0.1.0
 */

/**
 * 出力するファイル形式。csv / excel / parquet から選択してください。選択した形式に応じた拡張子が自動付与されます。
 */
export type ExportFileRequestBodyFormat = typeof ExportFileRequestBodyFormat[keyof typeof ExportFileRequestBodyFormat];


export const ExportFileRequestBodyFormat = {
  csv: 'csv',
  excel: 'excel',
  parquet: 'parquet',
} as const;